+++ client/src/router.ts
@@ -1,0 +1,4 @@
+import { createBrowserRouter } from 'react-router-dom'
+import { routes } from 'virtual:routes'
+
+export const router = createBrowserRouter(routes)
